requirements.txt
streamlit
weaviate_client